Reworked for merge request #27 review (part 2)
* Combined FluxLabelConverter and FluxConverter into 1 class
