+++ python/libneo/__init__.py
@@ -14,4 +14,3 @@ from .eqdsk_base import *
 from .convert_neo_in_to_fortran_namelist import *
 from .getHeaderDataVMEC import *
 from .flux_label_converter import *
-from .flux_converter import *
